backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
dj-database-url==3.0.1
psycopg[binary]==3.1.18
gunicorn==21.2.0
django-imagekit==4.3.3

Pillow==9.5.0
whitenoise==6.10.0
